Changed to long int

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,13 +8,13 @@
 int main(void)
 {
 	int i;
-	unsigned long long int fib1 = 1;
-	unsigned long long int fib2 = 2;
-	unsigned long long int future;
+	long int fib1 = 1;
+	long int fib2 = 2;
+	long int future;
 
 	for (i = 0; i < 50; i++)
 	{
-		printf("%llu", fib1);
+		printf("%ld", fib1);
 
 		future = fib1 + fib2;
 		fib1 = fib2;
